# Optimize detection of duplicate fields