--- a/internal/types/document_test.go
+++ b/internal/types/document_test.go
@@ -46,6 +46,7 @@ func TestDocument(t *testing.T) {
 
 		// to avoid {} != nil in tests
 		assert.Nil(t, must.NotFail(NewDocument()).fields)
+		assert.Nil(t, must.NotFail(NewDocument()).keys)
 
 		var doc Document
 		assert.Equal(t, 0, doc.Len())
@@ -99,6 +100,26 @@ func TestDocument(t *testing.T) {
 		a.Set("foo", "bar")
 		assert.NotEqual(t, a, b)
 		assert.Equal(t, int32(42), must.NotFail(b.Get("foo")))
+	})
+
+	t.Run("Duplicates", func(t *testing.T) {
+		t.Parallel()
+
+		doc, err := NewDocument("foo", int32(1), "foo", int32(2))
+		require.NoError(t, err)
+
+		assert.True(t, doc.Has("foo"))
+		assert.False(t, doc.Has("bar"))
+
+		assert.PanicsWithValue(t, "types.Document.Get: key is duplicated: foo", func() {
+			doc.Get("foo")
+		})
+		assert.PanicsWithValue(t, "types.Document.Set: key is duplicated: foo", func() {
+			doc.Set("foo", int32(3))
+		})
+		assert.PanicsWithValue(t, "types.Document.Remove: key is duplicated: foo", func() {
+			doc.Remove("foo")
+		})
 	})
 
 	t.Run("SortFieldsByKey", func(t *testing.T) {